Removes unnecessary NestJS entities and works around several Typeorm issues.
After review of the seed csv data, it appears to be a denormalized form
of transactions, people and a nondescript "entity".

This might "best" be represented as a join table, but that seems
unnecessary both for the assignment and given the lack of additional
information about the "entity" or person.

To better use this data in the denormalized form it is presented in,
only one entity representing the transaction itself is needed.

Typeorm has version and CLI conflicts in v0.3 and up, so I have reverted
it to 0.2.45 and updated the package.json alias.